pyproject.toml
[project]
name = "evaluatellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=4.29.0",
    "arize-phoenix-evals>=0.28.1",
    "arize-phoenix-otel>=0.13.0",
    "chonkie[all]>=1.1.2",
    "flagembedding>=1.3.5",
    "litellm[all]>=1.75.8",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "openai>=1.99.9",
    "openinference-instrumentation-litellm>=0.1.25",
    "openinference-instrumentation-openai>=0.1.31",
    "opentelemetry-sdk>=1.36.0",
    "phoenix>=0.9.1",
]
